requirements.txt
flask==2.3.3
flask-sqlalchemy==3.0.5
flask-migrate==4.0.4
flask-login==0.6.2
mysql-connector-python==8.1.0
pymysql==1.1.0
cryptography==41.0.3
flask-wtf==1.1.1
email-validator==2.0.0
python-dotenv==1.0.0
passlib==1.7.4
matplotlib==3.7.2
pandas==2.0.3
plotly==5.16.1
tabulate==0.9.0
termcolor==2.3.0
